Move module into its own file.
